rename: rename plugins modules, to coexists with other plugins
