+++ assistant_gettext.lua
