+++ assistant_mdparser.lua
